version 1.0.0 commit

+++ src/main/java/com/example/zegeju/Domain/Student/Student.java
@@ -13,7 +13,7 @@ public class Student {
     private String lastName;
     private String email;
     private String password;
-    private int phoneNumber;
+    private long phoneNumber;
     private int otp;
     private Date otpExpiry;
     private boolean verified;
